Intermediate broken reefactor
