+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export const BOARD_SIZE = 16;
