+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type Grid = number[][];
+export type Direction = "UP" | "DOWN" | "LEFT" | "RIGHT";
+export type Board = number[];
